Added function to split up DNS queries
This is a band-aid fix to reduce false positives that result from the script incorrectly generating alarms based on subdomains of legitimate domains. There are probably many ways to address this, but one I've used in the past is to break down the DNS query and then reassemble it so that it only includes the domain and TLD. The advantage to doing it like this is that it works independent of any legit_domains the user is interested in tracking ... except for country-code TLDs. They can probably be added, but right now they aren't handled at all. This won't cause problems unless someone adds them to the legit_domains variable.

Also added a $sub message to the notice to inform the analyst what the legitimate domain is supposed to be.